BattleHandle update
1.Update time calculate
2.Upload test part

+++ Assets/Scripts/GlobalDefine.cs
@@ -19,7 +19,7 @@ namespace GlobalDefine
 
     public class GameSetting
     {
-		public const float BATTLE_ROUND_TIME = 10f;		// 戰鬥一回合所需花費的時間(秒)
+		public const int BATTLE_ROUND_TIME = 10;		// 戰鬥一回合所需花費的時間(秒)
 
 		public const int NORMALSTAGE_MAX_FLOOR = 5000;	// 一般關卡最高樓層
